+++ .gitignore
@@ -2,6 +2,7 @@ coverage/
 dist/
 node_modules/
 packed/
+.parcel-cache/
 *.log
 
 # IDE
