+++ package.json
@@ -1,40 +1,29 @@
 {
   "name": "thread_downloader",
-  "version": "0.5.0",
+  "version": "0.6.0",
   "description": "",
-  "main": "threadDownloader.js",
   "scripts": {
-    "start": "nodemon src/index.ts",
+    "dev": "nodemon --watch src/index.ts --exec node --experimental-specifier-resolution=node --loader ts-node/esm src/index.ts",
     "build": "tsc",
-    "webpack": "webpack"
+    "bundle": "rollup -c"
   },
   "keywords": [],
   "author": "",
   "license": "MIT",
   "dependencies": {
     "cli-progress": "^3.11.2",
-    "prompts": "^2.4.2"
+    "prompts": "^2.4.2",
+    "rollup": "^3.2.3"
   },
-  "type": "module",
   "devDependencies": {
-    "@babel/core": "^7.19.6",
-    "@babel/preset-env": "^7.19.4",
+    "@rollup/plugin-commonjs": "^23.0.2",
+    "@rollup/plugin-node-resolve": "^15.0.1",
     "@types/cli-progress": "^3.11.0",
     "@types/prompts": "^2.4.1",
-    "babel-loader": "^8.2.5",
-    "buffer": "^6.0.3",
-    "events": "^3.3.0",
-    "https-browserify": "^1.0.0",
     "nodemon": "^2.0.19",
-    "process": "^0.11.10",
-    "querystring-es3": "^0.2.1",
-    "rimraf": "^3.0.2",
-    "stream-http": "^3.2.0",
+    "rollup-plugin-cleanup": "^3.2.1",
     "ts-node": "^10.8.2",
-    "typescript": "^4.7.4",
-    "url": "^0.11.0",
-    "webpack": "^5.74.0",
-    "webpack-cli": "^4.10.0",
-    "webpack-dev-server": "^4.11.1"
+    "tslib": "^2.4.0",
+    "typescript": "^4.7.4"
   }
 }
