+++ readme.md
@@ -1,5 +1,9 @@
 # 2ch Thread Downloader
 
+## What does it do:
+
+Allows you to download all files from specified thread.
+
 ## Usage:
 
 This tool requires [Node.js](https://nodejs.org/)
@@ -8,7 +12,7 @@ This tool requires [Node.js](https://nodejs.org/)
 node threadDownloader.cjs
 ```
 
-<img src="https://github.com/bloodvez/threadDownloader/blob/master/downloader.gif" alt="nut"/>
+<img src="https://github.com/bloodvez/threadDownloader/blob/master/src/img/downloader.gif" alt="nut"/>
 
 ---
 
@@ -18,16 +22,16 @@ node threadDownloader.cjs
 
 ```shell
 npm i
-npm run start
+npm run dev
 ```
 
 ---
 
 ### Building:
 1. Transpile with typescript.
-2. Bundle with webpack.
+2. Bundle with rollup.
 
 ```shell
 npm run build
-npm run webpack
+npm run bundle
 ```
