+++ src/progress-bar.ts
@@ -1,0 +1,5 @@
+import { Presets, SingleBar } from "cli-progress";
+export const progressBar = new SingleBar(
+  { format: " {bar} | {percentage}% | {filename} | {value}/{total}" },
+  Presets.legacy
+);
